--- docs/swagger/cosmos-routes.yml
@@ -1,57 +1,0 @@
-paths:
-  /cosmos:
-    get:
-      tags:
-        - 'cosmos'
-      summary: 'View the Cosmos network and RPC URL that gateway is configured to use'
-      description: 'The user can change this by editing src/chains/cosmos/cosmos.config.ts'
-      operationId: 'root'
-      produces:
-        - 'application/json'
-      responses:
-        '200':
-          description: 'Cosmos config'
-          schema:
-            $ref: '#/definitions/CosmosConfigResponse'
-
-  /cosmos/balances:
-    post:
-      tags:
-        - 'cosmos'
-      summary: 'Get the balances of a Cosmos wallet'
-      operationId: 'balances'
-      consumes:
-        - 'application/json'
-      produces:
-        - 'application/json'
-      parameters:
-        - in: 'body'
-          name: 'body'
-          required: true
-          schema:
-            $ref: '#/definitions/CosmosBalanceRequest'
-      responses:
-        '200':
-          schema:
-            $ref: '#/definitions/BalancesResponse'
-
-  /cosmos/poll:
-    post:
-      tags:
-        - 'cosmos'
-      summary: 'Poll the status of a Cosmos transaction'
-      operationId: 'poll'
-      consumes:
-        - 'application/json'
-      produces:
-        - 'application/json'
-      parameters:
-        - in: 'body'
-          name: 'body'
-          required: true
-          schema:
-            $ref: '#/definitions/CosmosPollRequest'
-      responses:
-        '200':
-          schema:
-            $ref: '#/definitions/CosmosPollResponse'
