+++ src/app.ts
@@ -2,7 +2,6 @@
 import express from 'express';
 import { Request, Response, NextFunction } from 'express';
 import { ConfigRoutes } from './services/config/config.routes';
-import { CosmosRoutes } from './chains/cosmos/cosmos.routes';
 import { WalletRoutes } from './services/wallet/wallet.routes';
 import { logger } from './services/logger';
 import { addHttps } from './https';
@@ -14,17 +13,13 @@ import {
 } from './services/error-handler';
 import { ConfigManagerV2 } from './services/config-manager-v2';
 import { SwaggerManager } from './services/swagger-manager';
-import { NetworkRoutes } from './network/network.routes';
 import { ConnectorsRoutes } from './connectors/connectors.routes';
-import { EVMRoutes } from './evm/evm.routes';
 import { AmmRoutes, AmmLiquidityRoutes, PerpAmmRoutes } from './amm/amm.routes';
-import { AlgorandRoutes } from './chains/algorand/algorand.routes';
-import { InjectiveRoutes } from './chains/injective/injective.routes';
-import { NearRoutes } from './chains/near/near.routes';
 import { CLOBRoutes, PerpClobRoutes } from './clob/clob.routes';
 
 import morgan from 'morgan';
 import swaggerUi from 'swagger-ui-express';
+import { ChainRoutes } from './chains/chain.routes';
 
 export const gatewayApp = express();
 
@@ -48,10 +43,7 @@ gatewayApp.use(
 
 // mount sub routers
 gatewayApp.use('/config', ConfigRoutes.router);
-gatewayApp.use('/network', NetworkRoutes.router);
-gatewayApp.use('/evm', EVMRoutes.router);
-gatewayApp.use('/algorand', AlgorandRoutes.router);
-gatewayApp.use('/injective', InjectiveRoutes.router);
+gatewayApp.use('/chain', ChainRoutes.router);
 gatewayApp.use('/connectors', ConnectorsRoutes.router);
 
 gatewayApp.use('/amm', AmmRoutes.router);
@@ -60,8 +52,6 @@ gatewayApp.use('/amm/liquidity', AmmLiquidityRoutes.router);
 gatewayApp.use('/wallet', WalletRoutes.router);
 gatewayApp.use('/clob', CLOBRoutes.router);
 gatewayApp.use('/clob/perp', PerpClobRoutes.router);
-gatewayApp.use('/cosmos', CosmosRoutes.router);
-gatewayApp.use('/near', NearRoutes.router);
 
 // a simple route to test that the server is running
 gatewayApp.get('/', (_req: Request, res: Response) => {
@@ -103,12 +93,7 @@ export const swaggerDocument = SwaggerManager.generateSwaggerJson(
     './docs/swagger/wallet-routes.yml',
     './docs/swagger/amm-routes.yml',
     './docs/swagger/amm-liquidity-routes.yml',
-    './docs/swagger/evm-routes.yml',
-    './docs/swagger/network-routes.yml',
-    './docs/swagger/algorand-routes.yml',
-    './docs/swagger/near-routes.yml',
-    './docs/swagger/cosmos-routes.yml',
-    './docs/swagger/injective-routes.yml',
+    './docs/swagger/chain-routes.yml',
   ]
 );
