+++ src/chains/algorand/algorand.controller.ts
@@ -3,22 +3,22 @@ import {
   AssetsRequest,
   AssetsResponse,
   OptInRequest,
-  OptInResponse,
   PollRequest,
   PollResponse,
 } from '../algorand/algorand.requests';
 import { Algorand } from './algorand';
-import {
-  BalanceRequest,
-  BalanceResponse,
-} from '../../network/network.requests';
-import { latency } from '../../services/base';
+import { BalanceRequest } from '../../network/network.requests';
 import {
   HttpException,
-  NETWORK_ERROR_CODE,
   TOKEN_NOT_SUPPORTED_ERROR_CODE,
   TOKEN_NOT_SUPPORTED_ERROR_MESSAGE,
 } from '../../services/error-handler';
+import {
+  validateAlgorandBalanceRequest,
+  validateAlgorandPollRequest,
+  validateAssetsRequest,
+  validateOptInRequest,
+} from './algorand.validators';
 
 async function getInitializedAlgorand(network: string): Promise<Algorand> {
   const algorand = Algorand.getInstance(network);
@@ -30,96 +30,86 @@ async function getInitializedAlgorand(network: string): Promise<Algorand> {
   return algorand;
 }
 
-export async function poll(
-  algorand: Algorand,
-  req: PollRequest
-): Promise<PollResponse> {
-  return await algorand.getTransaction(req.txHash);
-}
+export class AlgorandController {
+  static async poll(
+    algorand: Algorand,
+    req: PollRequest
+  ): Promise<PollResponse> {
+    validateAlgorandPollRequest(req);
 
-export async function balances(
-  chain: Algorand,
-  request: BalanceRequest
-): Promise<BalanceResponse> {
-  const initTime = Date.now();
-  const balances: Record<string, string> = {};
-
-  const account = await chain.getAccountFromAddress(request.address);
-
-  if (request.tokenSymbols.includes(chain.nativeTokenSymbol)) {
-    balances[chain.nativeTokenSymbol] = await chain.getNativeBalance(account);
+    return await algorand.getTransaction(req.txHash);
   }
 
-  for (const token of request.tokenSymbols) {
-    if (token === chain.nativeTokenSymbol) continue;
-    balances[token] = await chain.getAssetBalance(account, token);
+  static async balances(chain: Algorand, request: BalanceRequest) {
+    validateAlgorandBalanceRequest(request);
+
+    const balances: Record<string, string> = {};
+
+    const account = await chain.getAccountFromAddress(request.address);
+
+    if (request.tokenSymbols.includes(chain.nativeTokenSymbol)) {
+      balances[chain.nativeTokenSymbol] = await chain.getNativeBalance(account);
+    }
+
+    for (const token of request.tokenSymbols) {
+      if (token === chain.nativeTokenSymbol) continue;
+      balances[token] = await chain.getAssetBalance(account, token);
+    }
+
+    return {
+      balances: balances,
+    };
   }
 
-  return {
-    network: request.network,
-    timestamp: initTime,
-    latency: latency(initTime, Date.now()),
-    balances: balances,
-  };
-}
+  static async getTokens(
+    algorand: Algorand,
+    request: AssetsRequest
+  ): Promise<AssetsResponse> {
+    validateAssetsRequest(request);
 
-export async function getAssets(
-  request: AssetsRequest
-): Promise<AssetsResponse> {
-  if (request.network === undefined) {
-    throw new HttpException(
-      500,
-      'Missing network parameter',
-      NETWORK_ERROR_CODE
-    );
-  }
+    let assets: AlgorandAsset[] = [];
 
-  let assets: AlgorandAsset[] = [];
-  const algorand = await getInitializedAlgorand(request.network);
-
-  if (!request.assetSymbols) {
-    assets = algorand.storedAssetList;
-  } else {
-    let assetSymbols;
-    if (typeof request.assetSymbols === 'string') {
-      assetSymbols = [request.assetSymbols];
+    if (!request.assetSymbols) {
+      assets = algorand.storedAssetList;
     } else {
-      assetSymbols = request.assetSymbols;
+      let assetSymbols;
+      if (typeof request.assetSymbols === 'string') {
+        assetSymbols = [request.assetSymbols];
+      } else {
+        assetSymbols = request.assetSymbols;
+      }
+      for (const a of assetSymbols as []) {
+        assets.push(algorand.getAssetForSymbol(a) as AlgorandAsset);
+      }
     }
-    for (const a of assetSymbols as []) {
-      assets.push(algorand.getAssetForSymbol(a) as AlgorandAsset);
-    }
+
+    return {
+      assets: assets,
+    };
   }
 
-  return {
-    assets: assets,
-  };
-}
+  static async approve(request: OptInRequest) {
+    validateOptInRequest(request);
 
-export async function optIn(request: OptInRequest): Promise<OptInResponse> {
-  const initTime = Date.now();
+    const algorand = await getInitializedAlgorand(request.network);
+    const asset = algorand.getAssetForSymbol(request.assetSymbol);
 
-  const algorand = await getInitializedAlgorand(request.network);
-  const asset = algorand.getAssetForSymbol(request.assetSymbol);
+    if (asset === undefined) {
+      throw new HttpException(
+        500,
+        TOKEN_NOT_SUPPORTED_ERROR_MESSAGE + request.assetSymbol,
+        TOKEN_NOT_SUPPORTED_ERROR_CODE
+      );
+    }
 
-  if (asset === undefined) {
-    throw new HttpException(
-      500,
-      TOKEN_NOT_SUPPORTED_ERROR_MESSAGE + request.assetSymbol,
-      TOKEN_NOT_SUPPORTED_ERROR_CODE
+    const transactionResponse = await algorand.optIn(
+      request.address,
+      request.assetSymbol
     );
+
+    return {
+      assetId: (asset as AlgorandAsset).assetId,
+      transactionResponse: transactionResponse,
+    };
   }
-
-  const transactionResponse = await algorand.optIn(
-    request.address,
-    request.assetSymbol
-  );
-
-  return {
-    network: request.network,
-    timestamp: initTime,
-    latency: latency(initTime, Date.now()),
-    assetId: (asset as AlgorandAsset).assetId,
-    transactionResponse: transactionResponse,
-  };
 }
