+++ src/chains/polygon/polygon.ts
@@ -10,12 +10,14 @@ import { UniswapConfig } from '../../connectors/uniswap/uniswap.config';
 import { Ethereumish } from '../../services/common-interfaces';
 import { ConfigManagerV2 } from '../../services/config-manager-v2';
 import { OpenoceanConfig } from '../../connectors/openocean/openocean.config';
+import { EVMController } from '../ethereum/evm.controllers';
 
 export class Polygon extends EthereumBase implements Ethereumish {
   private static _instances: { [name: string]: Polygon };
   private _gasPrice: number;
   private _nativeTokenSymbol: string;
   private _chain: string;
+  public controller;
 
   private constructor(network: string) {
     const config = getPolygonConfig('polygon', network);
@@ -33,6 +35,7 @@ export class Polygon extends EthereumBase implements Ethereumish {
     this._chain = config.network.name;
     this._nativeTokenSymbol = config.nativeCurrencySymbol;
     this._gasPrice = config.manualGasPrice;
+    this.controller = EVMController;
   }
 
   public static getInstance(network: string): Polygon {
