+++ src/chains/xdc/xdc.base.ts
@@ -11,9 +11,9 @@ import { promises as fs } from 'fs';
 import path from 'path';
 import { rootPath } from '../../paths';
 import { TokenListType, TokenValue, walletPath } from '../../services/base';
-import { EVMNonceManager } from '../../evm/evm.nonce';
+import { EVMNonceManager } from '../ethereum/evm.nonce';
 import NodeCache from 'node-cache';
-import { EvmTxStorage } from '../../evm/evm.tx-storage';
+import { EvmTxStorage } from '../ethereum/evm.tx-storage';
 import fse from 'fs-extra';
 import { ConfigManagerCertPassphrase } from '../../services/config-manager-cert-passphrase';
 import { logger } from '../../services/logger';
