+++ src/network/network.controllers.ts
@@ -1,9 +1,4 @@
-import {
-  StatusRequest,
-  StatusResponse,
-  TokensRequest,
-  TokensResponse,
-} from './network.requests';
+import { StatusRequest, StatusResponse } from './network.requests';
 import { Avalanche } from '../chains/avalanche/avalanche';
 import { BinanceSmartChain } from '../chains/binance-smart-chain/binance-smart-chain';
 import { Ethereum } from '../chains/ethereum/ethereum';
@@ -16,10 +11,8 @@ import {
   UNKNOWN_CHAIN_ERROR_CODE,
   UNKNOWN_KNOWN_CHAIN_ERROR_MESSAGE,
 } from '../services/error-handler';
-import { EthereumBase, TokenInfo } from '../chains/ethereum/ethereum-base';
 import { Cronos } from '../chains/cronos/cronos';
 import { Near } from '../chains/near/near';
-import { Nearish, Xdcish } from '../services/common-interfaces';
 import { Algorand } from '../chains/algorand/algorand';
 import {
   getInitializedChain,
@@ -126,44 +119,4 @@ export async function getStatus(
   }
 
   return req.chain ? statuses[0] : statuses;
-}
-
-export async function getTokens(req: TokensRequest): Promise<TokensResponse> {
-  type connectionType = EthereumBase | Nearish | Injective | Xdcish;
-  let connection: connectionType;
-  let tokens: TokenInfo[] = [];
-
-  if (req.chain && req.network) {
-    try {
-      connection = (await getInitializedChain(
-        req.chain as string,
-        req.network as string
-      )) as connectionType;
-    } catch (e) {
-      if (e instanceof UnsupportedChainException) {
-        throw new HttpException(
-          500,
-          UNKNOWN_KNOWN_CHAIN_ERROR_MESSAGE(req.chain),
-          UNKNOWN_CHAIN_ERROR_CODE
-        );
-      }
-      throw e;
-    }
-  } else {
-    throw new HttpException(
-      500,
-      UNKNOWN_KNOWN_CHAIN_ERROR_MESSAGE(req.chain),
-      UNKNOWN_CHAIN_ERROR_CODE
-    );
-  }
-
-  if (!req.tokenSymbols) {
-    tokens = connection.storedTokenList;
-  } else {
-    for (const t of req.tokenSymbols as []) {
-      tokens.push(connection.getTokenForSymbol(t) as TokenInfo);
-    }
-  }
-
-  return { tokens };
 }
