+++ test-helpers/curl/requests/add_polygon_key.json
