+++ test-helpers/curl/requests/eth_allowances.json
@@ -1,7 +1,7 @@
 {
   "chain": "ethereum",
-  "network": "goerli",
-  "address": "$ETH_ADDRESS",
+  "network": "mainnet",
+  "address": "0x010216bB52E46807a07d0101Bb828bA547534F37",
   "spender": "uniswap",
   "tokenSymbols": ["DAI", "WETH"]
 }
