+++ test-helpers/curl/requests/injective_transfer_to_sub.json
@@ -2,7 +2,8 @@
   "chain": "injective",
   "network": "mainnet",
   "connector": "injective",
-  "address": "$INJECTIVE_SUBACCOUNTID",
+  "from": "$INJECTIVE_ACCOUNTID",
+  "to": "$INJECTIVE_SUBACCOUNTID",
   "amount": "0.123",
   "token": "USDT"
 }
