+++ test/chains/ethereum/evm.nonce.test.ts
@@ -12,7 +12,7 @@ import {
   SERVICE_UNITIALIZED_ERROR_CODE,
   SERVICE_UNITIALIZED_ERROR_MESSAGE,
 } from '../../../src/services/error-handler';
-import { EVMNonceManager } from '../../../src/evm/evm.nonce';
+import { EVMNonceManager } from '../../../src/chains/ethereum/evm.nonce';
 import { patch, unpatch } from '../../services/patch';
 
 import 'jest-extended';
